Interface,repository,controllerlar olusturuldu ve arasindaki baglantilar olusturuldu,hata alinan ve eksik kisimlar duzeltildi

+++ STS.Infrastructure/Repositories/StockRepository.cs
@@ -1,7 +1,9 @@
-﻿using STS.Domain.Entities;
-using STS.Infrastructure.Data;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
+using STS.Application.DTOs.Products;
+using STS.Application.DTOs.Stock;
 using STS.Application.IRepositories;
+using STS.Domain.Entities;
+using STS.Infrastructure.Data;
 
 namespace STS.Infrastructure.Repositories
 {
@@ -14,36 +16,93 @@ namespace STS.Infrastructure.Repositories
             _context = context;
         }
 
-        public async Task<Stock> GetByIdAsync(int id)
+        public async Task<StockReadDto> GetByIdAsync(int id)
         {
+
+
+            //STOCK ID GORE LİSTELE
             return await _context.Stocks
-                                 .Include(s => s.Product)
-                                 .FirstOrDefaultAsync(s => s.Id == id);
+                                 .Where(s => s.Id == id)
+                                 .Select(s => new StockReadDto {
+                                     Id = s.Id,
+                                     ProductId=s.ProductId,
+                                     ProductName = s.Product.Name,
+                                     Quantity=s.Quantity,
+                                     Store=s.Store
+
+                                 })
+                                 .FirstOrDefaultAsync();
         }
 
-        public async Task<IEnumerable<Stock>> GetAllAsync()
+        //STOCK LİSTELE
+        public async Task<IEnumerable<StockReadDto>> GetAllAsync()
         {
             return await _context.Stocks
-                                 .Include(s => s.Product)
+                                 .Select(s => new StockReadDto
+                                 { 
+                                     Id=s.Id,
+                                     ProductId=s.ProductId,
+                                     ProductName=s.Product.Name,
+                                     Quantity=s.Quantity,
+                                     Store=s.Store
+
+                                 })
                                  .ToListAsync();
         }
 
-        public async Task AddAsync(Stock stock)
+        //STOCK EKLEME
+        public async Task<StockReadDto> AddAsync(StockCreateDto dto)
         {
-            await _context.Stocks.AddAsync(stock);
+            var stock = new Stock
+            {
+                Id = dto.Id,
+                ProductId = dto.ProductId,
+                Store = dto.Store,
+                Quantity = dto.Quantity
+
+            };
+            //attach()
+            _context.Stocks.Add(stock);
             await _context.SaveChangesAsync();
+            return new StockReadDto
+            {
+                Id = stock.Id,
+                ProductId = stock.ProductId,
+                ProductName=stock.Product.Name,
+                Quantity=stock.Quantity,
+                Store=stock.Store
+
+            };
+
+
         }
 
-        public void Update(Stock stock)
+        public async Task UpdateAsync(StockUpdateDto dto)
         {
-            _context.Stocks.Update(stock);
-            _context.SaveChanges();
+            var stock = await _context.Stocks
+                        .Where(stock => stock.Id == dto.Id)
+                        .Select(stock => new Stock
+                        {
+                            Id = stock.Id,
+                            Store = stock.Store,
+                            Quantity = stock.Quantity
+
+
+                        }).FirstOrDefaultAsync();
+            if (stock == null)
+                return;
+            _context.Stocks.Attach(stock);
+
         }
 
-        public void Delete(Stock stock)
+        public async Task DeleteAsync(int id)
         {
-            _context.Stocks.Remove(stock);
-            _context.SaveChanges();
+            var stock = await _context.Stocks.FindAsync(id);
+            if(stock != null)
+            {
+                _context.Stocks.Remove(stock);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
